requirements.txt
# NeuroClaw_Bot — Python Dependencies

# ── Local Model Inference (CPU) ───────────────────────────────
llama-cpp-python==0.3.4
# GPU support: pip install llama-cpp-python==0.3.4 --extra-index-url
#              https://abetlen.github.io/llama-cpp-python/whl/cu121

# ── RAG / Embeddings ──────────────────────────────────────────
sentence-transformers==3.3.1
faiss-cpu>=1.7.4
numpy>=1.24

# ── Configuration ─────────────────────────────────────────────
pyyaml>=6.0

# ── Model Download ────────────────────────────────────────────
huggingface-hub>=0.23.0

# ── Optional: Telegram Remote Control ────────────────────────
# python-telegram-bot==21.6
